+++ server/pyproject.toml
@@ -7,7 +7,7 @@ requires-python = ">=3.12,<3.13"
 dependencies = [
     "alembic>=1.15.2",
     "openai>=1.99.3,<2",
-    "camel-ai==0.2.85a0",
+    "camel-ai==0.2.85",
     "pydantic[email]>=2.11.1",
     "click>=8.1.8",
     "fastapi>=0.115.12",
